+++ test/sql/install_psql.sql
@@ -16,8 +16,8 @@ end;
 -- check installation
 -----------------------------
 
--- check the emaj_param content
-SELECT param_value_text FROM emaj.emaj_param WHERE param_key = 'emaj_version';
+-- check the emaj_version_hist content
+select verh_version from emaj.emaj_version_hist;
 
 -- check history
 select hist_id, hist_function, hist_event, hist_object, hist_wording, hist_user from emaj.emaj_hist order by hist_id;